requirements.txt
Flask
Flask_CORS
pytest
black
requests
pyjwt
uuid
pymongo
Flask-PyMongo
python-dotenv
Flask-restful
Flask-apispec
apispec
marshmallow
pylint
Flask-socketio
Flask-restx
Flask-SQLAlchemy
Flask-Babel
mysqlclient 
python-dotenv